Live peripheral evaluation: sync DMA, instruction clock, TIM/WWDG/DCMI reads + demo wiring
- Rust core drains mem-to-mem DMA inline after EN store (NDTR/dst/TCIF
  visible immediately); live INSTRUCTION_COUNT in 16-inst chunks with
  count-less tick_peripherals() post-step tick
- TIM CNT advances on read; WWDG counter runs once CR-written regardless
  of WDGA (reset still needs WDGA), evaluated live on SR read; DCMI FNE
  bit + live-pull polled DR reads; u64 PWM duty math (no div-by-zero)
- Demo wiring: SPI3 flash (spi_tft/edge), I2C EEPROM (periph/edge), DCMI
  camera + boot feed (new/deep_periph), IRQ delivery (edge/periph)
- edge_test SPI clocking fixed in firmware (dummy-first reads) + rebuilt;
  verify_ethernet.sh paths fixed; cli.mjs debug cleaned; docs

+++ stm32-periph-wasm/src/peripherals/tim.rs
@@ -156,10 +156,14 @@ impl Timer {
             }
         }
 
-        // Update PWM duty based on CCR/ARR
+        // Update PWM duty based on CCR/ARR. u64 math: ARR==0xFFFFFFFF
+        // (reset default on 32-bit timers, or a transient config window with
+        // CC already enabled) would otherwise make (arr+1) wrap to zero and
+        // trap, and CCR*100 can overflow u32 for large CCR values. A full-
+        // range period yields ~0% duty, which is the honest answer.
         for ch in 0..4 {
             if self.ccer & (1 << (ch * 4)) != 0 && self.arr > 0 {
-                self.pwm_duty[ch] = self.ccr[ch] * 100 / (self.arr + 1);
+                self.pwm_duty[ch] = ((self.ccr[ch] as u64 * 100) / ((self.arr as u64) + 1)) as u32;
             }
         }
     }
